+++ ev.h
@@ -37,6 +37,8 @@
  * 1. fix: loop may wait infinite with EV_LOOP_MODE_ONCE if there are only endgame events
  * 2. fix: memory leak in create process
  * 3. fix: thread leak when process exit
+ * 4. fix: wrong format charster in test
+ * 
  * 
  * ## v0.0.8 (2023/04/07)
  * 
@@ -205,7 +207,7 @@ extern "C" {
 /**
  * @brief Development version.
  */
-#define EV_VERSION_PREREL           5
+#define EV_VERSION_PREREL           6
 
 /**
  * @brief Version calculate helper macro.