+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 1. fix: loop may wait infinite with EV_LOOP_MODE_ONCE if there are only endgame events
 2. fix: memory leak in create process
 3. fix: thread leak when process exit
+4. fix: wrong format charster in test
+
 
 ## v0.0.8 (2023/04/07)